+++ packages/components/src/components.d.ts
@@ -30,6 +30,11 @@ export namespace Components {
          */
         "active": boolean;
         /**
+          * The background color of the chip. The default #005655 HEX value refers to the --brand-colors-dark-pine-100 from MIDT
+          * @type {string}
+         */
+        "backgroundColor"?: string;
+        /**
           * The chip content that is displayed in the component.
           * @type {string}
          */
@@ -698,6 +703,11 @@ export namespace Components {
     }
     interface MiRouteInstructionsStep {
         /**
+          * The final step when the user has arrived to the destination location.
+          * @type {string}
+         */
+        "destinationLocation"?: string;
+        /**
           * The route context of previous step, if any.
           * @type {string}
          */
@@ -1266,6 +1276,11 @@ declare namespace LocalJSX {
           * @type {boolean}
          */
         "active"?: boolean;
+        /**
+          * The background color of the chip. The default #005655 HEX value refers to the --brand-colors-dark-pine-100 from MIDT
+          * @type {string}
+         */
+        "backgroundColor"?: string;
         /**
           * The chip content that is displayed in the component.
           * @type {string}
@@ -1943,6 +1958,11 @@ declare namespace LocalJSX {
         "unit"?: UnitSystem;
     }
     interface MiRouteInstructionsStep {
+        /**
+          * The final step when the user has arrived to the destination location.
+          * @type {string}
+         */
+        "destinationLocation"?: string;
         /**
           * The route context of previous step, if any.
           * @type {string}