+++ packages/components/src/components/route-instructions-step/route-instructions-step.tsx
@@ -56,6 +56,12 @@ export class RouteInstructionsStep implements ComponentInterface {
     @Prop() fromTransitStop: string;
 
     /**
+     * The final step when the user has arrived to the destination location.
+     * @type {string}
+     */
+    @Prop() destinationLocation?: string;
+
+    /**
      * Set imperial or metric as default unit system. Default is Metric unless the browser is running US English. In that case Imperial.
      * @type {UnitSystem}
      */
@@ -133,6 +139,8 @@ export class RouteInstructionsStep implements ComponentInterface {
                 return this.renderBicyclingStep();
             case 'TRANSIT':
                 return this.renderTransitStep();
+            case 'DESTINATION':
+                return this.renderDestinationStep();
             default:
                 return <div>Unknown travel mode: {this.stepData.travel_mode}</div>;
         }
@@ -161,6 +169,9 @@ export class RouteInstructionsStep implements ComponentInterface {
                 break;
             case 'TRANSIT':
                 travelModeIcon = this.getTransitVehicleIconName();
+                break;
+            case 'DESTINATION':
+                travelModeIcon = 'logo';
                 break;
         }
 
@@ -261,19 +272,17 @@ export class RouteInstructionsStep implements ComponentInterface {
      */
     renderDrivingStep(): JSX.Element {
         return <div class="step" onClick={e => this.stepClickHandler(e)}>
-            {this.isInternetExplorer ? null :
-                <span class="step__action-icon step__action-icon--circled">
-                    <mi-icon icon-name="car"></mi-icon>
-                </span>
-            }
-            <h3 part="step-heading" class="step__heading">{this.getStepHeading()}</h3>
-            {this.renderTravelMode()}
             <div part="step-description" class="step__description">
-                {this.translationsData.drive}<br />
                 {this.renderDistance()}
             </div>
-            {this.renderToggleButton()}
-            {this.renderSubsteps()}
+            <div class="step__info">
+                {this.isInternetExplorer ? null :
+                    <span class="step__action-icon step__action-icon--circled">
+                        <mi-icon icon-name="car"></mi-icon>
+                    </span>
+                }
+                <div part="step-heading" class="step__heading">{this.getStepHeading()}</div>
+            </div>
         </div>;
     }
 
@@ -286,7 +295,7 @@ export class RouteInstructionsStep implements ComponentInterface {
         const defaultHeadings = {
             'driving': this.translationsData.drive,
             'walking': this.translationsData.walk,
-            'bicycling': this.translationsData.bike
+            'bicycling': this.translationsData.bike,
         };
 
         return this.stepData.steps[0]?.instructions ?
@@ -398,19 +407,38 @@ export class RouteInstructionsStep implements ComponentInterface {
      */
     renderBicyclingStep(): JSX.Element {
         return <div class="step" onClick={e => this.stepClickHandler(e)}>
-            {this.isInternetExplorer ? null :
-                <span class="step__action-icon step__action-icon--circled">
-                    <mi-icon icon-name="bike"></mi-icon>
-                </span>
-            }
-            <h3 part="step-heading" class="step__heading">{this.getStepHeading()}</h3>
-            {this.renderTravelMode()}
             <div part="step-description" class="step__description">
-                {this.translationsData.bike}<br />
                 {this.renderDistance()}
             </div>
-            {this.renderToggleButton()}
-            {this.renderSubsteps()}
+            <div class="step__info">
+                {this.isInternetExplorer ? null :
+                    <span class="step__action-icon step__action-icon--circled">
+                        <mi-icon icon-name="bike"></mi-icon>
+                    </span>
+                }
+                <div part="step-heading" class="step__heading">{this.getStepHeading()}</div>
+            </div>
+        </div>;
+    }
+
+    /**
+     * Render the destination step.
+     *
+     * @returns {JSX.Element}
+     */
+    renderDestinationStep(): JSX.Element {
+        return <div class="step" onClick={e => this.stepClickHandler(e)}>
+            <div class="step__info destination">
+                <div class="step__destination">
+                    {this.isInternetExplorer ? null :
+                        <span class="step__action-icon step__action-icon--circled">
+                            <mi-icon icon-name="logo"></mi-icon>
+                        </span>
+                    }
+                    <div part="step-heading" class="step__heading">{this.translationsData.destination}</div>
+                </div>
+                <div part="step-location" class="step__location">{this.destinationLocation}</div>
+            </div>
         </div>;
     }