+++ packages/map-template/src/atoms/filteredLocationsByExternalIDState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const filteredLocationsByExternalIDState = atom({
+    key: 'filteredLocationsByExternalID',
+    default: undefined
+});
+
+export default filteredLocationsByExternalIDState;