+++ packages/map-template/src/atoms/tileStyleState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const tileStyleState = atom({
+    key: 'tileStyle',
+    default: null
+});
+
+export default tileStyleState;