+++ packages/map-template/src/constants/travelModes.js
@@ -1,0 +1,5 @@
+export const travelModes = {
+    WALKING: 'WALKING',
+    DRIVING: 'DRIVING',
+    BICYCLING: 'BICYCLING'
+};